Use faketime's -f option in tests
Without this change, faketime will pass the provided timestamp
directly through to the system `date -d`. On GNU systems this is
apparenty fine, but FreeBSD `date` doesn't like this and will throw an
error since apparently on those systems `-d` has something to do with
setting the system Daylight Savings Time (see date(1) for more).

So instead we use `faketime -f`, which passes the provided string
directly through to libfaketime. This means that the parsing will be
consistent cross-platform, regardless of the underlying `date` binary.